Update
point 생성 안하는대 왜 안하는지 알아보고 적용-> BattleManager Pool에 겹치는게 있었음 버그 해결!,
PlayerMovement, EnemyBullet,LivingObject, Bullet,Blaster,Laser, TakeDamge 처리 방식 변경, 상속 확정,
LivingObject에 spriteRenderer 보호 수준 변경

+++ Framework/Assets/Script/Manager/GameManager.cs
@@ -851,8 +851,8 @@ public class GameManager : MonoBehaviour
         playerData.position = new Vector3(posX, posY, posZ);
 
         // 체력 및 기타 플레이어 데이터 로드
-        playerData.health = PlayerPrefs.GetInt("PlayerHealth", 20); // 기본 값 6
-        playerData.Maxhealth = PlayerPrefs.GetInt("PlayerMaxHealth", 20);
+        playerData.health = PlayerPrefs.GetFloat("PlayerHealth", 20f); // 기본 값 6
+        playerData.Maxhealth = PlayerPrefs.GetFloat("PlayerMaxHealth", 20f);
         playerData.player_Name = PlayerPrefs.GetString("PlayerName", playerData.player_Name);
 
         // 인벤토리 아이템 로드
